refactor(abstract-utxo): split transaction.ts into multiple files
Issue: BTC-1450
