+++ modules/abstract-utxo/src/transaction/index.ts
@@ -1,0 +1,3 @@
+export * from './recipient';
+export * from './explainTransaction';
+export * from './fetchInputs';
